chat_project/requirements.txt
django==2
psycopg2
easy-thumbnails
ipython
simplejson
djangorestframework
djangorestframework-jwt
markdown
django-filter
django-bootstrap3
channels
asgi_redis
